Split HTTP functionality of `@typespec/rest` to `@typespec/http`
