+++ packages/http/.c8rc.json
@@ -1,0 +1,3 @@
+{
+  "reporter": ["cobertura", "json", "text"]
+}
