+++ packages/http/CHANGELOG.json
@@ -1,0 +1,4 @@
+{
+  "name": "@typespec/http",
+  "entries": []
+}
