+++ packages/http/mocha.reporter.config.json
@@ -1,0 +1,3 @@
+{
+  "reporterEnabled": "spec, mocha-junit-reporter"
+}
