+++ packages/http/src/index.ts
@@ -8,3 +8,4 @@ export * from "./parameters.js";
 export * from "./responses.js";
 export * from "./route.js";
 export * from "./types.js";
+export * from "./validate.js";
