+++ packages/openapi3/package.json
@@ -13,9 +13,7 @@
   "bugs": {
     "url": "https://github.com/Microsoft/typespec/issues"
   },
-  "keywords": [
-    "typespec"
-  ],
+  "keywords": ["typespec"],
   "type": "module",
   "main": "dist/src/index.js",
   "exports": {
@@ -24,12 +22,8 @@
   },
   "typesVersions": {
     "*": {
-      "*": [
-        "./dist/src/index.d.ts"
-      ],
-      "testing": [
-        "./dist/src/testing/index.d.ts"
-      ]
+      "*": ["./dist/src/index.d.ts"],
+      "testing": ["./dist/src/testing/index.d.ts"]
     }
   },
   "tspMain": "dist/src/index.js",
@@ -46,17 +40,14 @@
     "lint": "eslint . --ext .ts --max-warnings=0",
     "lint:fix": "eslint . --fix --ext .ts"
   },
-  "files": [
-    "lib/*.tsp",
-    "dist/**",
-    "!dist/test/**"
-  ],
+  "files": ["lib/*.tsp", "dist/**", "!dist/test/**"],
   "dependencies": {
     "js-yaml": "~4.1.0"
   },
   "peerDependencies": {
     "@typespec/versioning": "~0.40.0",
     "@typespec/compiler": "~0.40.0",
+    "@typespec/http": "~0.40.0",
     "@typespec/rest": "~0.40.0",
     "@typespec/openapi": "~0.40.0"
   },
@@ -65,6 +56,7 @@
     "@types/node": "~18.11.9",
     "@types/js-yaml": "~4.0.1",
     "@typespec/compiler": "~0.40.0",
+    "@typespec/http": "~0.40.0",
     "@typespec/rest": "~0.40.0",
     "@typespec/openapi": "~0.40.0",
     "@typespec/versioning": "~0.40.0",
