+++ packages/rest/lib/resource.tsp
@@ -1,4 +1,4 @@
-import "./http.tsp";
+import "@typespec/http";
 import "../dist/src/index.js";
 import "../dist/src/internal-decorators.js";
