+++ packages/rest/package.json
@@ -13,28 +13,18 @@
   "bugs": {
     "url": "https://github.com/Microsoft/typespec/issues"
   },
-  "keywords": [
-    "typespec"
-  ],
+  "keywords": ["typespec"],
   "type": "module",
   "main": "dist/src/index.js",
   "tspMain": "lib/rest.tsp",
   "exports": {
     ".": "./dist/src/index.js",
-    "./http": "./dist/src/http/index.js",
     "./testing": "./dist/src/testing/index.js"
   },
   "typesVersions": {
     "*": {
-      "*": [
-        "./dist/src/index.d.ts"
-      ],
-      "http": [
-        "./dist/src/http/index.d.ts"
-      ],
-      "testing": [
-        "./dist/src/testing/index.d.ts"
-      ]
+      "*": ["./dist/src/index.d.ts"],
+      "testing": ["./dist/src/testing/index.d.ts"]
     }
   },
   "engines": {
@@ -50,11 +40,7 @@
     "lint": "eslint . --ext .ts --max-warnings=0",
     "lint:fix": "eslint . --fix --ext .ts"
   },
-  "files": [
-    "lib/*.tsp",
-    "dist/**",
-    "!dist/test/**"
-  ],
+  "files": ["lib/*.tsp", "dist/**", "!dist/test/**"],
   "peerDependencies": {
     "@typespec/compiler": "~0.40.0"
   },
@@ -62,6 +48,7 @@
     "@types/mocha": "~10.0.0",
     "@types/node": "~18.11.9",
     "@typespec/compiler": "~0.40.0",
+    "@typespec/http": "~0.40.0",
     "@typespec/eslint-config-typespec": "~0.5.0",
     "@typespec/library-linter": "~0.40.0",
     "@typespec/eslint-plugin": "~0.40.0",
