+++ packages/rest/src/index.ts
@@ -1,5 +1,4 @@
 export const namespace = "TypeSpec.Rest";
-export * as http from "./http/index.js";
 export * from "./resource.js";
 export * from "./rest.js";
 export * from "./validate.js";
