+++ packages/samples/package.json
@@ -14,10 +14,7 @@
   "bugs": {
     "url": "https://github.com/Microsoft/typespec/issues"
   },
-  "keywords": [
-    "typespec",
-    "cli"
-  ],
+  "keywords": ["typespec", "cli"],
   "type": "module",
   "engines": {
     "node": ">=16.0.0"
@@ -28,14 +25,11 @@
     "watch": "tsc -p . --watch",
     "regen-samples": "node scripts/regen-samples.js"
   },
-  "files": [
-    "lib/*.tsp",
-    "dist/**",
-    "!dist/test/**"
-  ],
+  "files": ["lib/*.tsp", "dist/**", "!dist/test/**"],
   "dependencies": {
     "@typespec/versioning": "~0.40.0",
     "@typespec/compiler": "~0.40.0",
+    "@typespec/http": "~0.40.0",
     "@typespec/rest": "~0.40.0",
     "@typespec/openapi": "~0.40.0",
     "@typespec/openapi3": "~0.40.0",
