+++ packages/samples/visibility/visibility.tsp
@@ -1,6 +1,8 @@
+import "@typespec/http";
 import "@typespec/rest";
 
 using TypeSpec.Http;
+using TypeSpec.Rest;
 
 @service({
   title: "Visibility sample",
